Add ajax attachment upload

+++ src/main/java/ua/ukma/nc/dao/ProjectAttachmentDao.java
@@ -16,7 +16,7 @@ public interface ProjectAttachmentDao {
 
     List<ProjectAttachment> getAll();
 
-    int createProjectAttachment(ProjectAttachment projectAttachment);
+    Long createProjectAttachment(ProjectAttachment projectAttachment);
 
 	List<ProjectAttachment> getAllById(Long id_project);
 }
